+++ Js2IL.Tests/Integration/DependencyPlannerTests.cs
@@ -25,10 +25,9 @@ public class DependencyPlannerTests
         builder.Build(module);
         var symbolTable = module.SymbolTable ?? throw new InvalidOperationException("Symbol table was not built.");
 
-        var logger = new TestLogger();
         var options = new CompilerOptions { Verbose = true };
         var registry = new CallableRegistry();
-        var coordinator = new TwoPhaseCompilationCoordinator(logger, options, registry);
+        var coordinator = new TwoPhaseCompilationCoordinator(options, registry);
         coordinator.RunPhase1Discovery(symbolTable);
 
         return (symbolTable, coordinator);